Defining the 'Player class'
Name, Position, and GemCount are the three attributes of the class. The constructor initializes Name and Position while setting GemCount to 0. You must also make sure that your code defines the Position class, commonly known as the struct, in order to finish this class. Position must be properly declared and imported if it is an enum or another class.